Clean program.cs and split routes into separate files each
